+++ domain/src/main/java/com/saegil/domain/repository/FeedRepository.kt
@@ -8,7 +8,7 @@ interface FeedRepository {
 
     fun getFeeds(
         query: String?,
-        organization: Int?,
+        organization: Long?,
     ): Flow<PagingData<Notice>>
 
 }
